chore: switch test runner from multirust to rustup.
I *should* switch to the common script repo, but there's an extra script
in here I don't want to have to mess with.  Suppose I could make a branch
or something, but I'll deal with that later.

Also, funny reason for this: rustup's multirust alias refuses to accept
1.11.0 as an installed toolchain.  rustup itself is fine.  I have *no*
idea why.